Update `### RedHat 6 and 7` in README
I was a little confused when I first read the README because it didn't mention RedHat 8, so this PR fixes that. It might also make sense to clarify line 5 because the role doesn't actually do anything on RedHat-based systems.

Also, thank you for all for all of your opensource work, YouTube channel, and Ansible books! I really appreciate it.